requirements.txt
streamlit
pdfplumber
rank_bm25


beautifulsoup4
requests
python-dotenv
# local-only (comment these out for Cloud unless you truly need them)
selenium
webdriver-manager
undetected-chromedriver

numpy<2
torch>=2.1.0,<2.2
sentence-transformers>=2.5.1,<3
chromadb>=0.4.24

undetected_chromedriver

